requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
python-dotenv==0.21.0
sqlparse==0.4.3
